Mappers

+++ src/main/java/com/carlosh/nesemulator/Bus.java
@@ -9,7 +9,7 @@ public class Bus {
   private PPU ppu;
   private CPU cpu;
   private ROM rom;
-  private int clockCounter = 0;
+  private long clockCounter = 0;
   public int[] controller = new int[2];
   private int[] controllerState = new int[2];
   int[] ram = new int[2048];
@@ -25,14 +25,11 @@ public class Bus {
     if (address >= 0x0000 && address <= 0x1FFF) {
       ram[address & 0x07FF] = data;
     } else if (address >= 0x2000 && address <= 0x3FFF) {
-      //System.out.println("Writing");
-      //System.out.println("Writing ppu.cpuWrite: " + (address & 0x0007) + " " + data);
       ppu.cpuWrite(address & 0x0007, data);
     } else if (address == 0x4014) {
       directAddress = data << 8;
       dma = true;
     } else if (address >= 0x4016 && address <= 0x4017) {
-      //System.out.println("BBBB");
       controllerState[address & 0x0001] = controller[address & 0x0001];
     }
   }
@@ -42,23 +39,17 @@ public class Bus {
 
     int romData = rom.cpuRead(address, readOnly);
     if (romData != -2) {
-      //System.out.println("Reading romData: " + romData);
       return romData;
     }
     if (address >= 0x0000 && address <= 0x1FFF) {
-      //System.out.println("Reading ram: " + ram[address & 0x07FF]);
       data = ram[address & 0x07FF];
-      //System.out.println("Reading ram 2: " + data);
       return data;
     } else if (address >= 0x2000 && address <= 0x3FFF) {
       data = ppu.cpuRead(address & 0x0007, readOnly);
-      //System.out.println("Reading ram 3: " + data + " Address: " + (address & 0x0007));
       return data;
     } else if (address >= 0x4016 && address <= 0x4017) {
-      //System.out.println("AAAAAA");
       data = (controllerState[address & 0x0001] & 0x80) > 0 ? 1 : 0;
       controllerState[address & 0x0001] <<= 1;
-      //System.out.println("Reading controller: " + data);
       return data;
     }
     return 0x00;
